# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-floatadd",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute `a + b` as an `f64`.",
        "description": "Input validation is delegated to `srvcs-isnumber` over HTTP (the single\nsource of truth for \"is this a number\"), once per operand. If that\ndependency is unreachable, this service reports itself degraded rather than\nguessing.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SumResponse"
                }
              }
            }
          },
          "422": {
            "description": "an operand is not a number"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "a",
          "b"
        ],
        "properties": {
          "a": {
            "type": "object"
          },
          "b": {
            "type": "object"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "SumResponse": {
        "type": "object",
        "required": [
          "a",
          "b",
          "result"
        ],
        "properties": {
          "a": {
            "type": "object"
          },
          "b": {
            "type": "object"
          },
          "result": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}